фиксы по требованиям Вовы
