+++ app/build.gradle.kts
@@ -44,12 +44,11 @@ dependencies {
 
     implementation("androidx.core:core-ktx:1.9.0")
     implementation("androidx.appcompat:appcompat:1.6.1")
-    //implementation("com.google.android.material:material:1.10.0")
     implementation("com.google.android.material:material:1.8.0")
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-
+    implementation("androidx.preference:preference:1.2.1")
     implementation("org.osmdroid:osmdroid-android:6.1.17")
 }
