+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:name=".PinCreateActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".MapActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
