+++ app/src/main/java/ru/nsu/ccfit/tsd/pinmap/PinMarker.kt
@@ -1,7 +1,7 @@
 package ru.nsu.ccfit.tsd.pinmap
 
 import android.content.Context
-import android.content.Intent
+import android.widget.Toast
 import org.osmdroid.views.MapView
 import org.osmdroid.views.overlay.Marker
 
@@ -11,11 +11,11 @@ class PinMarker : Marker {
     constructor(mapView: MapView, ctx: Context) : super(mapView, ctx)
 
     override fun onMarkerClickDefault(marker: Marker?, mapView: MapView?): Boolean {
-        /* этот листенер вызывает окно редактирования маркера по нажатию */
+        // этот листенер вызывает окно редактирования маркера по нажатию
         val context = mapView!!.context
-        val intent = Intent(context, PinEditActivity::class.java)
-        context.startActivity(intent)
-        setInfoWindow(null) /* это чтобы не показывать InfoWindow при нажатии на маркер */
+        // TODO: Вызывать PinConstructorActivity
+        Toast.makeText(context, "TODO: Вызывать PinConstructorActivity", Toast.LENGTH_SHORT).show()
+        setInfoWindow(null) // это чтобы не показывать InfoWindow при нажатии на маркер
         return super.onMarkerClickDefault(marker, mapView)
     }
